udpate sim tests for new org
